cluster_setup: improve the reliability of the cleanup
It's possible that we would get a GeneratorExit exception in the main()
funciton because of @inlineCallbacks. GeneratorExit is not a subclass of
Exception, so it's better to catch BaseException here (or to use the
dangerous naked 'except').

Also, we now destroy the cluster if there is any problem with provisioning
the nodes (installing the Flocker) or with configuring the Flocker cluster
on top of them.